text-analysis-tool/requirements.txt
openai
google-generativeai
mistralai
transformers
gradio
pandas
matplotlib
textblob
langdetect
